# packages/api-client/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/devices/register": {
      "post": {
        "operationId": "DevicesController_register",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterDeviceDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegisteredDeviceDto_Output"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Devices"
        ]
      }
    },
    "/api/devices/current/staff": {
      "get": {
        "operationId": "DevicesController_roster",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RosterEntryDto_Output"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "device": []
          }
        ],
        "tags": [
          "Devices"
        ]
      }
    },
...
    "/api/staff": {
      "post": {
        "operationId": "StaffController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateStaffDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StaffCreatedDto_Output"
                }
              }
            }
          }
        },
        "security": [
          {
            "device": []
          },
          {
            "bearer": []
          }
        ],
        "tags": [
          "Staff"
        ]
      }
    },
    "/api/staff/{id}/pin": {
      "put": {
        "operationId": "StaffController_setPin",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetPinDto"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "device": []
          },
          {
            "bearer": []
          }
        ],
        "tags": [
          "Staff"
        ]
      }
    },
    "/api/sync/push": {
      "post": {
        "operationId": "SyncController_push",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PushDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PushResponseDto_Output"
                }
              }
            }
          }
        },
        "security": [
          {
            "device": []
          }
        ],
        "tags": [
          "Sync"
        ]
      }
    },
    "/api/sync/pull": {
      "get": {
        "operationId": "SyncController_pull",
        "parameters": [
          {
            "name": "cursor",
            "required": false,
            "in": "query",
            "schema": {
              "pattern": "^\\d+$",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PullResponseDto_Output"
                }
              }
            }
          }
        },
        "security": [
          {
            "device": []
          }
        ],
        "tags": [
          "Sync"
        ]
      }
    }
  },
  "info": {
    "title": "Plateraa API",
    "description": "",
    "version": "1",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "device": {
        "type": "apiKey",
        "in": "header",
        "name": "x-device-token"
      },
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "RegisterDeviceDto": {
        "type": "object",
        "properties": {
          "tenantId": {
            "type": "string",
            "format": "ulid",
            "pattern": "^[0-7][0-9A-HJKMNP-TV-Za-hjkmnp-tv-z]{25}$"
          },
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 60
          },
          "platform": {
            "type": "string",
            "maxLength": 40
          },
          "appVersion": {
            "type": "string",
            "maxLength": 40
          }
        },
        "required": [
          "tenantId",
          "name"
        ]
      },
      "RegisteredDeviceDto_Output": {
        "type": "object",
        "properties": {
          "deviceToken": {
            "type": "string"
          },
          "device": {
            "type": "object",
            "properties": {
              "id": {
                "type": "string"
              },
              "tenantId": {
                "type": "string"
              },
              "locationId": {
                "type": "string"
              },
              "name": {
                "type": "string"
              },
              "code": {
                "type": "string"
              }
            },
            "required": [
              "id",
              "tenantId",
              "locationId",
              "name",
              "code"
            ],
            "additionalProperties": false
          }
        },
        "required": [
          "deviceToken",
          "device"
        ],
        "additionalProperties": false
      },
      "RosterEntryDto_Output": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "displayName": {
            "type": "string"
          },
          "role": {
            "type": "string",
            "enum": [
              "OWNER",
              "MANAGER",
              "STAFF",
              "RIDER"
            ]
          },
          "pinVerifier": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "id",
          "displayName",
          "role",
          "pinVerifier"
        ],
        "additionalProperties": false
      },
...
      "CreateStaffDto": {
        "type": "object",
        "properties": {
          "displayName": {
            "type": "string",
            "minLength": 1,
            "maxLength": 60
          },
          "role": {
            "type": "string",
            "enum": [
              "MANAGER",
              "STAFF",
              "RIDER"
            ]
          },
          "pin": {
            "type": "string"
          }
        },
        "required": [
          "displayName",
          "role",
          "pin"
        ]
      },
      "StaffCreatedDto_Output": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "displayName": {
            "type": "string"
          },
          "role": {
            "type": "string",
            "enum": [
              "OWNER",
              "MANAGER",
              "STAFF",
              "RIDER"
            ]
          }
        },
        "required": [
          "id",
          "displayName",
          "role"
        ],
        "additionalProperties": false
      },
      "SetPinDto": {
        "type": "object",
        "properties": {
          "pin": {
            "type": "string"
          }
        },
        "required": [
          "pin"
        ]
      },
...
      "PushResponseDto_Output": {
        "type": "object",
        "properties": {
          "results": {
            "type": "array",
            "items": {
              "anyOf": [
                {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "status": {
                      "type": "string",
                      "enum": [
                        "APPLIED"
                      ]
                    },
                    "result": {
                      "type": "object",
                      "additionalProperties": {}
                    }
                  },
                  "required": [
                    "id",
                    "status",
                    "result"
                  ],
                  "additionalProperties": false
                },
                {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "status": {
                      "type": "string",
                      "enum": [
                        "REJECTED",
                        "RETRY"
                      ]
                    },
                    "error": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "string"
                        },
                        "message": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "code",
                        "message"
                      ],
                      "additionalProperties": false
                    }
                  },
                  "required": [
                    "id",
                    "status",
                    "error"
                  ],
                  "additionalProperties": false
                }
              ]
            }
          }
        },
        "required": [
          "results"
        ],
        "additionalProperties": false
      },
      "PullResponseDto_Output": {
        "type": "object",
        "properties": {
          "cursor": {
            "type": "string"
          },
          "changes": {
            "type": "object",
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "object",
                "additionalProperties": {}
              }
            }
          }
        },
        "required": [
          "cursor",
          "changes"
        ],
        "additionalProperties": false
      }
    }
  }
}